Bring dropwizard files inside nervewsgi collector
